MobileNav made into its own component. CSS class report renamed desktopContainer.
